chore: some fixes

+++ Lecture10/src/migrations/1707600000000-AddImageFields.ts
@@ -2,27 +2,40 @@ import { MigrationInterface, QueryRunner, TableColumn } from 'typeorm';
 
 export class AddImageFields1707600000000 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.addColumn(
-      'users',
-      new TableColumn({
-        name: 'avatarUrl',
-        type: 'varchar',
-        isNullable: true,
-      }),
-    );
+    const hasAvatarUrl = await queryRunner.hasColumn('users', 'avatarUrl');
+    if (!hasAvatarUrl) {
+      await queryRunner.addColumn(
+        'users',
+        new TableColumn({
+          name: 'avatarUrl',
+          type: 'varchar',
+          isNullable: true,
+        }),
+      );
+    }
 
-    await queryRunner.addColumn(
-      'products',
-      new TableColumn({
-        name: 'imageUrl',
-        type: 'varchar',
-        isNullable: true,
-      }),
-    );
+    const hasImageUrl = await queryRunner.hasColumn('products', 'imageUrl');
+    if (!hasImageUrl) {
+      await queryRunner.addColumn(
+        'products',
+        new TableColumn({
+          name: 'imageUrl',
+          type: 'varchar',
+          isNullable: true,
+        }),
+      );
+    }
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.dropColumn('users', 'avatarUrl');
-    await queryRunner.dropColumn('products', 'imageUrl');
+    const hasAvatarUrl = await queryRunner.hasColumn('users', 'avatarUrl');
+    if (hasAvatarUrl) {
+      await queryRunner.dropColumn('users', 'avatarUrl');
+    }
+
+    const hasImageUrl = await queryRunner.hasColumn('products', 'imageUrl');
+    if (hasImageUrl) {
+      await queryRunner.dropColumn('products', 'imageUrl');
+    }
   }
 }
